+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Note that since we don't clearly distinguish between a public and private interf
 ## [Unreleased]
 
 - Add occlusion color parameter
+- Fix issue with outlines and orthographic camera
 
 ## [v3.31.4] - 2023-02-24